renamed RunInCloud -> RunInCloudDialog
