+++ src/components/RunInCloudDialog/states/index.tsx
@@ -40,5 +40,3 @@ export function RunInCloudStates({ state }: RunInCloudStatesProps) {
       return exhaustive(state)
   }
 }
-
-export * from './types'
